deleted the "fallback"-code, which was one of the reasons for slowing down fluxbox in utf8-locales:
if no fontset is found up to this point its because of
  a) no such font
  b) not supported locale

if we try our best and enhance the basename-list and add some
more generic patterns the xserver will give us a valid
fontset .. which is good coz we have something to work with
but which is BAD coz we return a valid fontset as if we 
have found the font the user wants .. so, every fallback
in higher level code will fail coz we find a valid fontset by any
meaning here. if no fontset can be found, the "fixed"-fontset should
be there already.